parser/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2==2.9.10
requests==2.32.3
lxml==5.3.1
pydantic==2.10.6
email_validator==2.2.0
python_dotenv==1.0.1